Warn user about insufficient window size
Comment on terminal saying window size is insufficient for given text.
I have noticed a very weird issue with printing on new line in the
terminal.
Firstly, regular print doesn't work while curses is running.
But sys.stdout.write does.
Secondly, none of these print or write functions work as expected
when a newline is printed. The terminal prompt username basically
overrides the print. But when no newline is printed, both of these
work well. But of course, the output is not well formatted then.
This is something to look for.