Squash multiple whitespaces
Multiple whitespace combinations (spaces, tabs, newlines)
together must be converted to a single space. This is really
important as user might pass an invalid text file with newlines
in it for example.
We currently do not support newlines, nor am I planning
to add it as it just increases confusion. Best way to retain
compatibility would be to squash all whitespaces together.