refacto: add intermediate function for show_history
The `show_history` method aim to print history logs of the users.
It has a single `number_record` parameter used to select the number of
record to print.

The major part of this function is used to extract the log data from the
`.mitype_history.csv` csv file. The other part is used to print this
extract.

This commit separate this two features in order to clean the code and
ease testing.